--- a/x-pack/apm-server/sampling/eventstorage/storage_test.go
+++ b/x-pack/apm-server/sampling/eventstorage/storage_test.go
@@ -5,7 +5,6 @@
 package eventstorage_test
 
 import (
-	"encoding/json"
 	"fmt"
 	"testing"
 	"time"
@@ -13,8 +12,10 @@ import (
 	"github.com/dgraph-io/badger/v2"
 	"github.com/gofrs/uuid"
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/testing/protocmp"
 
 	"github.com/elastic/apm-data/model/modelpb"
@@ -37,7 +38,7 @@ func TestWriteEvents(t *testing.T) {
 
 func testWriteEvents(t *testing.T, numSpans int) {
 	db := newBadgerDB(t, badgerOptions)
-	store := eventstorage.New(db, eventstorage.JSONCodec{})
+	store := eventstorage.New(db, eventstorage.ProtobufCodec{})
 	readWriter := store.NewShardedReadWriter()
 	defer readWriter.Close()
 
@@ -67,7 +68,13 @@ func testWriteEvents(t *testing.T, numSpans int) {
 	// We can read our writes without flushing.
 	var batch modelpb.Batch
 	assert.NoError(t, readWriter.ReadTraceEvents(traceID, &batch))
-	assert.ElementsMatch(t, append(spanEvents, &transaction), batch)
+	spanEvents = append(spanEvents, &transaction)
+	assert.Empty(t, cmp.Diff(modelpb.Batch(spanEvents), batch,
+		cmpopts.SortSlices(func(e1 *modelpb.APMEvent, e2 *modelpb.APMEvent) bool {
+			return e1.GetSpan().GetId() < e2.GetSpan().GetId()
+		}),
+		protocmp.Transform()),
+	)
 
 	// Flush in order for the writes to be visible to other readers.
 	assert.NoError(t, readWriter.Flush(wOpts.StorageLimitInBytes))
@@ -99,7 +106,7 @@ func testWriteEvents(t *testing.T, numSpans int) {
 			var event modelpb.APMEvent
 			require.Equal(t, "e", string(item.UserMeta()))
 			assert.NoError(t, item.Value(func(data []byte) error {
-				return json.Unmarshal(data, &event)
+				return proto.Unmarshal(data, &event)
 			}))
 			recorded = append(recorded, &event)
 		}
@@ -110,7 +117,7 @@ func testWriteEvents(t *testing.T, numSpans int) {
 
 func TestWriteTraceSampled(t *testing.T) {
 	db := newBadgerDB(t, badgerOptions)
-	store := eventstorage.New(db, eventstorage.JSONCodec{})
+	store := eventstorage.New(db, eventstorage.ProtobufCodec{})
 	readWriter := store.NewShardedReadWriter()
 	defer readWriter.Close()
 	wOpts := eventstorage.WriterOpts{
@@ -163,18 +170,24 @@ func TestWriteTraceSampled(t *testing.T) {
 
 func TestReadTraceEvents(t *testing.T) {
 	db := newBadgerDB(t, badgerOptions)
-	store := eventstorage.New(db, eventstorage.JSONCodec{})
+	store := eventstorage.New(db, eventstorage.ProtobufCodec{})
 
 	traceID := [...]byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16}
 	require.NoError(t, db.Update(func(txn *badger.Txn) error {
 		key := append(traceID[:], ":12345678"...)
-		value := []byte(`{"transaction":{"name":"transaction"}}`)
+		value, err := proto.Marshal(&modelpb.APMEvent{Transaction: &modelpb.Transaction{Name: "transaction"}})
+		if err != nil {
+			return err
+		}
 		if err := txn.SetEntry(badger.NewEntry(key, value).WithMeta('e')); err != nil {
 			return err
 		}
 
 		key = append(traceID[:], ":87654321"...)
-		value = []byte(`{"span":{"name":"span"}}`)
+		value, err = proto.Marshal(&modelpb.APMEvent{Span: &modelpb.Span{Name: "span"}})
+		if err != nil {
+			return err
+		}
 		if err := txn.SetEntry(badger.NewEntry(key, value).WithMeta('e')); err != nil {
 			return err
 		}
@@ -182,14 +195,14 @@ func TestReadTraceEvents(t *testing.T) {
 		// Write an entry with the trace ID as a prefix, but with no
 		// proceeding colon, causing it to be ignored.
 		key = append(traceID[:], "nocolon"...)
-		value = []byte(`not-json`)
+		value = []byte(`not-protobuf`)
 		if err := txn.SetEntry(badger.NewEntry(key, value).WithMeta('e')); err != nil {
 			return err
 		}
 
 		// Write an entry with an unknown meta value. It will be ignored.
 		key = append(traceID[:], ":11111111"...)
-		value = []byte(`not-json`)
+		value = []byte(`not-protobuf`)
 		if err := txn.SetEntry(badger.NewEntry(key, value).WithMeta('?')); err != nil {
 			return err
 		}
@@ -201,15 +214,15 @@ func TestReadTraceEvents(t *testing.T) {
 
 	var events modelpb.Batch
 	assert.NoError(t, reader.ReadTraceEvents(string(traceID[:]), &events))
-	assert.Equal(t, modelpb.Batch{
+	assert.Empty(t, cmp.Diff(modelpb.Batch{
 		{Transaction: &modelpb.Transaction{Name: "transaction"}},
 		{Span: &modelpb.Span{Name: "span"}},
-	}, events)
+	}, events, protocmp.Transform()))
 }
 
 func TestReadTraceEventsDecodeError(t *testing.T) {
 	db := newBadgerDB(t, badgerOptions)
-	store := eventstorage.New(db, eventstorage.JSONCodec{})
+	store := eventstorage.New(db, eventstorage.ProtobufCodec{})
 
 	traceID := [...]byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16}
 	require.NoError(t, db.Update(func(txn *badger.Txn) error {
@@ -231,7 +244,7 @@ func TestReadTraceEventsDecodeError(t *testing.T) {
 
 func TestIsTraceSampled(t *testing.T) {
 	db := newBadgerDB(t, badgerOptions)
-	store := eventstorage.New(db, eventstorage.JSONCodec{})
+	store := eventstorage.New(db, eventstorage.ProtobufCodec{})
 
 	require.NoError(t, db.Update(func(txn *badger.Txn) error {
 		if err := txn.SetEntry(badger.NewEntry([]byte("sampled_trace_id"), nil).WithMeta('s')); err != nil {
@@ -276,7 +289,7 @@ func TestStorageLimit(t *testing.T) {
 	db = newBadgerDB(t, opts)
 	lsm, vlog := db.Size()
 
-	store := eventstorage.New(db, eventstorage.JSONCodec{})
+	store := eventstorage.New(db, eventstorage.ProtobufCodec{})
 	readWriter := store.NewReadWriter()
 	defer readWriter.Close()